2.2.2: read & write minor updates

+++ include/base/filesystem.h
@@ -53,8 +53,8 @@ typedef off_t offset_t;
      int  filesystem_open2     ( char *pathname, int flags, mode_t mode ) ;
      int  filesystem_close     ( int fd ) ;
 
-     long filesystem_read      ( int read_fd2,  void *buffer, int buffer_size ) ;
-     long filesystem_write     ( int write_fd2, void *buffer, int num_bytes_to_write ) ;
+     ssize_t filesystem_read   ( int read_fd2,  void *buffer, size_t buffer_size ) ;
+     ssize_t filesystem_write  ( int write_fd2, void *buffer, size_t num_bytes_to_write ) ;
 
      int  filesystem_rename    ( char *old_pathname, char *new_pathname ) ;
